eventengine: split engine.go into queue and evaluation
engine.go crossed the soft 1500 LOC [WATCH] floor in #6810 (1477 ->
1540). That crossing was recorded in docs/refactoring-audit-accepted.txt
as a deliberate deferral rather than a dismissal, because #6810 was a
lost-remediation fix whose entire diff lived in enqueue and HandleEvent
— exactly the code a queue split would move — and a verbatim move
carrying the fix inside it makes "is this line new, or moved?"
unanswerable on a change whose whole point is a subtle ordering property
between latch-arming and queue admission.

This is that deferral's own PR, and it is a VERBATIM MOVE. No behaviour
change, so the diff reads as a move:

  engine.go    943  construction, Apply, HandleEvent, the action worker,
                    runAction/applyOnce, cooldown arming, accessors
  queue.go     215  admission: enqueue, supersede, releaseEdgeLatch,
                    actionQueueDepth — a bounded-channel admission
                    policy with its own concurrency contract (#5062
                    producer serialization, #5853 early dedup, #2869
                    FIFO, #6810 the admission verdict)
  evaluate.go  427  evaluation: evaluateEvent, withinMatches,
                    pruneWindow, attributesMatch, classifyPlan and
                    helpers — the within-clause semantics (#3751,
                    #3756, #7223, #4423) read as one subject and were
                    separated from themselves by the worker code

THE LOCK ORDER IS WHAT THIS SPLIT PUTS AT RISK, so it is now stated in
all three files rather than once. e.mu (evaluation and runtime state)
and enqueueMu (admission) are never nested in either order: enqueue runs
after evaluateEvent has released e.mu, which is precisely what lets
releaseEdgeLatch take e.mu at all, since #6810's latch rollback would
deadlock against a held enqueueMu. Before this change both locks were
visible on one screen. They are not any more, and that is the whole
reason the invariant is repeated instead of cross-referenced once.

Proven rather than asserted, because "verbatim" is a claim:

  - all 10 moved functions byte-identical to their originals AND absent
    from engine.go (no copy left behind, no reformatting)
  - declaration census 48 before, 48 after, empty diff, no duplicates
  - the 16-file pkg/eventengine suite green with ZERO test edits, which
    is the issue's own guardrail: a pure move that needs a test changed
    is not a pure move

Pruned the [WATCH] entry from docs/refactoring-audit-accepted.txt — it
is a decision record, not state, and the decision has now been carried
out. pkg/eventengine/README.md gains the file layout and the lock-order
note, and its one stale withinMatches reference now points at
evaluate.go.

Validation: go build ./... clean; go test -count=1 ./pkg/eventengine/
green; full ./... gate in the PR body.
