+++ pkg/eventengine/engine.go
@@ -59,7 +59,6 @@ import (
 	"regexp"
 	"sort"
 	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -90,15 +89,6 @@ type CommitFn func(ctx context.Context, comment string) (*config.Config, error)
 
 // Minimum time between successive triggers of the same policy.
 const policyCooldown = 30 * time.Second
-
-// actionQueueDepth bounds the worker's pending-action channel. Dedup-by-policy
-// (a newer trigger supersedes an older queued one) keeps at most one pending
-// action per policy, so this is also an upper bound on distinct policies with
-// a remediation in flight while the config lock is held. #5853: the dedup runs
-// on EVERY enqueue (not only when the channel is full), so a burst from ONE
-// policy occupies at most ONE slot — it can never fill the queue with redundant
-// duplicates and starve an unrelated policy's remediation into a queue-full drop.
-const actionQueueDepth = 64
 
 // Backoff schedule for a held config lock (#2157). The worker retries an
 // action whose EnterConfigure fails with ErrConfigLocked until the deadline,
@@ -274,6 +264,13 @@ type Engine struct {
 	// never nested in either order. Every channel op performed under enqueueMu is
 	// a non-blocking select-with-default, and the consumer never holds enqueueMu,
 	// so the queue always drains and no producer can block indefinitely.
+	//
+	// #7636: the two sides of this invariant now live in DIFFERENT FILES —
+	// enqueueMu's users in queue.go, e.mu's in evaluate.go and here. That makes
+	// the rule easier to break by accident, because the locks are no longer
+	// visible on one screen, so it is restated at the top of both files. If you
+	// are adding a call from one to the other, that is the invariant you are
+	// about to test.
 	enqueueMu sync.Mutex
 
 	// lifeCtx is the engine-lifetime context threaded into the remediation
@@ -629,188 +626,6 @@ func (e *Engine) Stats() Stats {
 	}
 }
 
-// enqueue adds an action to the bounded worker queue with dedup-by-policy: a
-// newer trigger of the same policy supersedes an older queued one (there is no
-// value in applying a stale remediation twice), which also bounds the queue to
-// one pending action per policy. If the queue is full of OTHER policies'
-// actions, the new action is dropped (counted) rather than blocking the caller
-// goroutine (#2157 bounded queue).
-//
-// #5853: the dedup runs on EVERY enqueue via supersede, not only when the
-// channel is full. The pre-#5853 fast path did an UNCONDITIONAL send first and
-// only deduped in the full/`default` branch, so while the worker was blocked
-// behind the config lock a burst from ONE policy filled all 64 slots with
-// redundant duplicates (later discarded by cooldown/staleness) and the next
-// remediation for an UNRELATED policy was dropped queue-full. Draining and
-// replacing the same-policy entry up front keeps at most one pending action per
-// policy, so a same-policy burst occupies a single slot and leaves the rest free
-// for other policies. supersede is non-blocking (select-with-default drain +
-// refill), so this never blocks the caller even mid-shutdown (e.actions is never
-// closed); a genuine full-of-other-policies queue still drops the new action
-// (counted) via supersede returning false.
-//
-// The whole body runs under enqueueMu (#5062) so concurrent producers cannot
-// interleave: supersede's drain+refill must be atomic w.r.t. other producers,
-// otherwise a second producer could take a slot supersede freed while draining
-// and force supersede to drop an already-accepted survivor. See the enqueueMu
-// field comment for the lock-ordering rationale.
-// enqueue returns whether the action was ADMITTED — i.e. whether an equivalent
-// action for this policy is now queued for the worker. It returns false only
-// when nothing will run: a genuine capacity drop, or a shutdown fast-exit.
-//
-// #6810: the verdict exists because the caller has already armed the edge latch
-// for this crossing by the time it gets here. Before this, enqueue returned
-// nothing, so a dropped action was indistinguishable from an admitted one at
-// the call site and the latch stayed armed over a remediation that never ran.
-// A superseded placement still counts as admitted: the newer equivalent action
-// IS queued, which is exactly what the latch is asserting.
-func (e *Engine) enqueue(a plannedAction) bool {
-	e.enqueueMu.Lock()
-	defer e.enqueueMu.Unlock()
-	// Fast exit during shutdown: don't churn the queue for an action the worker
-	// will never apply. supersede would otherwise still succeed (the channel is
-	// unbounded-in-shutdown only in that it is never closed), but there is no
-	// consumer left to drain it.
-	select {
-	case <-e.stopCh:
-		return false
-	default:
-	}
-	if e.supersede(a) {
-		return true
-	}
-	// supersede could not place `a`: the queue is full of OTHER policies'
-	// actions and there was no same-policy entry to evict. This is the only
-	// genuine capacity drop (an unrelated policy really did fill the queue).
-	e.counters.droppedQueueFull.Add(1)
-	slog.Warn("event-options: action queue full, dropping remediation",
-		"policy", a.policyName)
-	return false
-}
-
-// releaseEdgeLatch clears the #3756 edge latch that evaluateEvent armed for one
-// crossing, when the action that crossing authorized was never admitted (#6810).
-//
-// evaluateEvent arms the latch under e.mu and returns; HandleEvent classifies
-// and enqueues afterwards, outside that lock. If the queue is full of OTHER
-// policies' actions the remediation is dropped — and because withinMatches
-// suppresses every later at/above-threshold event while the latch is armed, and
-// only re-arms when a clause's count falls BELOW its threshold, a transient
-// queue saturation permanently consumed the crossing. For a sustained fault the
-// level never drops, so the configured remediation simply never ran.
-//
-// Rolling the latch back restores the invariant the latch is supposed to
-// express: "this crossing already fired". A crossing whose action was dropped
-// did not fire.
-//
-// authRev is the policy's semantic revision as of the evaluate that armed the
-// latch, and the guard is the same ABA defence armCooldown uses (#5311): if a
-// successor generation was installed (or the policy removed) while this action
-// was being classified and rejected, its latch state belongs to the successor
-// and must not be cleared by a predecessor's failure. Identity check and clear
-// happen in ONE critical section under e.mu so a concurrent Apply cannot swap
-// the runtime between them.
-//
-// Lock order is safe by construction: the caller has already released
-// enqueueMu (enqueue returns before this runs), so e.mu is never taken while
-// enqueueMu is held.
-//
-// A concurrent event that lands between the drop and this rollback is still
-// suppressed by the armed latch — the window is one classify+enqueue — so the
-// crossing costs at most a delay to the next event rather than being consumed
-// outright, which is the whole of the defect.
-func (e *Engine) releaseEdgeLatch(name, eventName, authRev string) {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	if e.semRev[name] != authRev {
-		// Successor generation installed, or the policy removed, while this
-		// action was rejected. Its latch is not ours to clear.
-		return
-	}
-	if rt := e.runtime[name]; rt != nil {
-		rt.onLatched[eventName] = false
-	}
-}
-
-// supersede non-blockingly rebuilds the queue, replacing any existing
-// same-policy action with a, and returns true if a was placed. It is the SOLE
-// enqueue path (#5853): every enqueue drains the buffered queue, drops any
-// existing same-policy entry (benign dedup, counted as superseded), re-enqueues
-// the surviving OTHER-policy actions in FIFO order, and places a at the tail. It
-// returns false only when the queue is full of OTHER policies and a could not be
-// placed (a genuine capacity drop the caller counts as droppedQueueFull).
-//
-// CALLER MUST HOLD enqueueMu (#5062). That is what makes the drain->re-enqueue
-// atomic w.r.t. other producers: while this runs, no other enqueue/supersede can
-// take a slot the drain just freed, so every surviving other-policy action is
-// re-enqueued exactly once in FIFO order (a survivor can never be dropped). The
-// only concurrent actor is the consumer (actionWorker), which just REMOVES
-// items, so the drain sees at most the buffered entries and the loop is bounded
-// with no need to guard against a producer refilling underneath it.
-func (e *Engine) supersede(a plannedAction) bool {
-	drained := make([]plannedAction, 0, actionQueueDepth)
-	replaced := false
-	// Drain whatever is currently buffered.
-	for {
-		select {
-		case old := <-e.actions:
-			e.counters.queueDepth.Add(-1)
-			if old.policyName == a.policyName {
-				// Drop the stale same-policy action; it is superseded by the
-				// newer trigger a. This is a benign dedup — the newer equivalent
-				// action still runs — NOT a capacity loss, so count it as
-				// superseded rather than droppedQueueFull (#5853). Inflating the
-				// alert-worthy queue_full metric on every same-policy burst
-				// (which the early dedup makes routine) would mask real capacity
-				// drops.
-				e.counters.superseded.Add(1)
-				continue
-			}
-			drained = append(drained, old)
-		default:
-			goto refill
-		}
-	}
-refill:
-	// Test-only seam (#5062): the drain->re-enqueue boundary. In production this
-	// is nil. A concurrency test sets it to drive a second producer into the
-	// freed slots here and prove enqueueMu serializes it (the survivors are
-	// preserved regardless of what the injected producer does).
-	if e.afterDrainFn != nil {
-		e.afterDrainFn()
-	}
-	// Re-enqueue the surviving other-policy actions in their original FIFO
-	// order, then place the new (superseding) action at the TAIL (#2869).
-	// Prepending `a` would jump it ahead of every already-queued action of
-	// OTHER policies, converting the documented FIFO queue into LIFO for the
-	// newest arrival and starving older queued remediations under sustained
-	// event frequency. Supersede must only drop/replace the stale SAME-policy
-	// entry (done in the drain loop above); it must not reorder unrelated
-	// policies relative to the order their events were observed.
-	all := append(drained, a)
-	for _, item := range all {
-		select {
-		case e.actions <- item:
-			e.counters.queueDepth.Add(1)
-			if item.policyName == a.policyName {
-				replaced = true
-			}
-		default:
-			// Still no room (lost the race to another producer, or the queue
-			// is full of unrelated policies and there was no stale same-policy
-			// entry to evict). Count the loss for any SURVIVOR we could not
-			// re-place — but NOT for the new action `a` itself: supersede
-			// returns false in that case and enqueue owns `a`'s single
-			// queue_full count + warn. Counting it here too would
-			// double-increment xpf_event_actions_dropped_total (#2869).
-			if item.policyName != a.policyName {
-				e.counters.droppedQueueFull.Add(1)
-			}
-		}
-	}
-	return replaced
-}
-
 // actionWorker is the single goroutine that applies remediation actions. It
 // serializes all configure/commit on behalf of the engine (removing the
 // cross-probe race) and retries a held config lock with bounded backoff.
@@ -1103,46 +918,6 @@ func (e *Engine) staleReason(a plannedAction) string {
 	return ""
 }
 
-// classifyPlan pre-parses a policy's ThenCommands into a typed plan WITHOUT
-// touching the candidate (#2139 step 1). An unknown command prefix, an
-// unparseable set, or an unparseable delete makes the WHOLE plan invalid
-// (ok=false) — the cheapest place to reject (no lock taken, no queue slot).
-func (e *Engine) classifyPlan(pol *config.EventPolicy) ([]plannedOp, bool) {
-	ops := make([]plannedOp, 0, len(pol.ThenCommands))
-	for _, cmd := range pol.ThenCommands {
-		cmd = strings.TrimSpace(cmd)
-		if cmd == "" {
-			continue
-		}
-		switch {
-		case strings.HasPrefix(cmd, "set "):
-			input := strings.TrimPrefix(cmd, "set ")
-			// Validate it parses now so a typo rejects the batch before any
-			// candidate mutation.
-			if _, err := config.ParseSetCommand("set " + input); err != nil {
-				slog.Warn("event-options: set parse failed (batch rejected)",
-					"policy", pol.Name, "cmd", cmd, "err", err)
-				return nil, false
-			}
-			ops = append(ops, plannedOp{setInput: input, raw: cmd})
-		case strings.HasPrefix(cmd, "delete "):
-			input := strings.TrimPrefix(cmd, "delete ")
-			path, err := config.ParseSetCommand("set " + input)
-			if err != nil {
-				slog.Warn("event-options: delete parse failed (batch rejected)",
-					"policy", pol.Name, "cmd", cmd, "err", err)
-				return nil, false
-			}
-			ops = append(ops, plannedOp{isDelete: true, delPath: path, raw: cmd})
-		default:
-			slog.Warn("event-options: unsupported command type (batch rejected)",
-				"policy", pol.Name, "cmd", cmd)
-			return nil, false
-		}
-	}
-	return ops, true
-}
-
 // armCooldown records a successful-commit timestamp for the policy under lock
 // (#2140 / #2157). Done on the worker after commit so a dropped/rejected
 // action does not consume the cooldown.
@@ -1172,369 +947,4 @@ func (e *Engine) armCooldown(name, authRev string) {
 	if rt := e.runtime[name]; rt != nil {
 		rt.lastTrigger = e.now()
 	}
-}
-
-// evaluateEvent checks policies under lock and returns any that should trigger.
-// It records the event in the sliding window (pruning on append so a
-// perpetually-cooldown-suppressed event can never grow unbounded — SMR finding
-// 1) and CHECKS (does not arm) the cooldown; the cooldown is armed by the
-// worker on a successful commit.
-func (e *Engine) evaluateEvent(ev rpm.Event) []triggeredPolicy {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-
-	var triggered []triggeredPolicy
-	now := e.now()
-	// #4423 M6/M5: scan only the policies that list this event (index built at
-	// Apply), not every policy on every event. The index guarantees the event
-	// matches, so the per-policy eventMatches scan is gone; the remaining
-	// per-matching-policy attributes/within work is inherent and bounded by the
-	// pruned window.
-	for _, pol := range e.eventIndex[ev.Name] {
-		if pol == nil {
-			continue
-		}
-
-		if !e.attributesMatch(pol, ev) {
-			continue
-		}
-
-		rt := e.runtime[pol.Name]
-		if rt == nil {
-			// Defensive: Apply always seeds runtime for every policy, but a
-			// policy could be evaluated before its first Apply in a test.
-			rt = newPolicyRuntime()
-			e.runtime[pol.Name] = rt
-		}
-
-		// Record this event, pruning the window on append so a suppressed
-		// event cannot grow it unbounded (SMR finding 1).
-		rt.windows[ev.Name] = append(rt.windows[ev.Name], now)
-		e.pruneWindow(pol, ev.Name, now)
-
-		if !e.withinMatches(pol, rt, ev.Name, now) {
-			continue
-		}
-
-		// Cooldown CHECK (armed on successful commit, not here): suppress a
-		// re-trigger within the cooldown window so a flapping probe does not
-		// flood the queue.
-		if !rt.lastTrigger.IsZero() && now.Sub(rt.lastTrigger) < policyCooldown {
-			continue
-		}
-
-		// #3756 M1: arm the edge latch for this crossing so a sustained
-		// above-threshold level does not re-fire every cooldown. Set only
-		// AFTER the cooldown check passed — a cooldown-suppressed crossing is
-		// NOT treated as already-fired, so the next event past the cooldown
-		// still fires. Cleared by withinMatches when the count drops back
-		// below the trigger-on threshold (re-arm).
-		if policyHasTriggerOn(pol) {
-			rt.onLatched[ev.Name] = true
-		}
-
-		slog.Info("event-options policy triggered",
-			"policy", pol.Name,
-			"event", ev.Name,
-			"test-owner", ev.TestOwner,
-			"test-name", ev.TestName)
-
-		// Capture the policy's semantic revision under the same lock so the
-		// enqueued action can be revalidated against it before commit (#3750).
-		triggered = append(triggered, triggeredPolicy{pol: pol, semRev: e.semRev[pol.Name]})
-	}
-	return triggered
-}
-
-// attributesMatch checks if the event attributes match the policy's filters.
-// Format: "ping_test_failed.test-owner matches <pattern>".
-//
-// Junos `attributes-match ... matches ...` semantics are a REGEX match, not
-// literal equality (this was a parity defect, #2008 M7). The operator's
-// pattern is treated as an RE2 regular expression compiled once at Apply()
-// time and cached here.
-//
-// #2141 FAIL-CLOSED: a malformed line (no " matches " / no ".") or an unknown
-// <field> name is rejected at commit (config.ValidateEventAttributesMatchStrict),
-// so it can only reach here on the legacy lenient-LOAD path (a config persisted
-// by an older binary). In that case the matcher fails CLOSED — the policy does
-// NOT trigger — rather than dropping the constraint and broadening the policy
-// (the dangerous direction per #2124). This is a behavior change for legacy
-// typo'd configs: they now STOP firing rather than over-fire, and are surfaced
-// by the boot-time lenient-compile warning plus the AttributesInvalid counter.
-func (e *Engine) attributesMatch(pol *config.EventPolicy, ev rpm.Event) bool {
-	for _, attr := range pol.AttributesMatch {
-		eventName, field, pattern, ok := config.ParseEventAttributesMatch(attr)
-		if !ok {
-			// Malformed line that slipped through a lenient load: fail closed.
-			e.flagAttributesInvalid(pol.Name, attr, "malformed match expression")
-			return false
-		}
-
-		// #3753: the constraint is scoped to a specific event name. Only apply
-		// it when the CURRENT event IS that event; a constraint written for a
-		// DIFFERENT event of a multi-event policy must NOT gate this event
-		// (before this the prefix was dropped, so `event_a.test-owner ...`
-		// incorrectly gated event_b too — and vice-versa). Commit-time
-		// validation guarantees eventName is one of the policy's events, so an
-		// out-of-scope prefix here can only be a legacy lenient-load config.
-		if eventName != ev.Name {
-			continue
-		}
-
-		if !config.EventAttributesFieldKnown(field) {
-			// Unknown field that slipped through a lenient load: fail closed.
-			e.flagAttributesInvalid(pol.Name, attr, "unknown field")
-			return false
-		}
-
-		var value string
-		switch field {
-		case "test-owner":
-			value = ev.TestOwner
-		case "test-name":
-			value = ev.TestName
-		case "target":
-			value = ev.Target
-		case "routing-instance":
-			value = ev.RoutingInstance
-		case "destination-interface":
-			value = ev.DestinationInterface
-		default:
-			// Known to config but not yet resolvable on rpm.Event. The SSOT
-			// and this switch are kept identical by a drift-guard test, so
-			// this is unreachable; fail closed if it is ever hit.
-			e.flagAttributesInvalid(pol.Name, attr, "unresolvable field")
-			return false
-		}
-
-		re := e.regexCache[pattern]
-		if re == nil {
-			// Defensive: pattern was not cached (commit validation should
-			// have rejected an uncompilable pattern, and Apply caches every
-			// valid one). Compile on demand; an uncompilable pattern fails
-			// the constraint CLOSED (the policy does not fire).
-			compiled, err := regexp.Compile(pattern)
-			if err != nil {
-				e.flagAttributesInvalid(pol.Name, attr, "uncompilable pattern")
-				return false
-			}
-			re = compiled
-			// #4423 M10: back-fill the cache so a pattern that slipped past the
-			// Apply-time build (legacy lenient-load path) is compiled ONCE, not
-			// on every event. Production callers reach here under e.mu, which
-			// also guards the Apply-time rebuild, so the write is serialized.
-			if e.regexCache == nil {
-				e.regexCache = make(map[string]*regexp.Regexp)
-			}
-			e.regexCache[pattern] = re
-		}
-
-		if !re.MatchString(value) {
-			return false
-		}
-	}
-	return true
-}
-
-// flagAttributesInvalid bumps the counter and emits a throttled warning when a
-// malformed/unknown attributes-match line is hit at runtime (#2141). The
-// throttle is PER POLICY (#4423 M11): a flapping bad line on one policy must
-// not swallow the first warning about a distinct bad line on another policy.
-func (e *Engine) flagAttributesInvalid(policy, attr, reason string) {
-	e.counters.attributesInvalid.Add(1)
-	now := e.now().UnixNano()
-	e.invalidWarnMu.Lock()
-	if e.invalidWarnAt == nil {
-		e.invalidWarnAt = make(map[string]int64)
-	}
-	last := e.invalidWarnAt[policy]
-	warn := last == 0 || now-last >= int64(10*time.Second)
-	if warn {
-		e.invalidWarnAt[policy] = now
-	}
-	e.invalidWarnMu.Unlock()
-	if warn {
-		slog.Warn("event-options: attributes-match invalid, policy fails closed (will not fire)",
-			"policy", policy, "line", attr, "reason", reason)
-	}
-}
-
-// policyHasTriggerOn reports whether the policy carries any `within { trigger
-// on N }` clause. Only such a policy participates in the #3756 M1 edge latch;
-// a no-within policy (or a `trigger until` policy) is unaffected.
-func policyHasTriggerOn(pol *config.EventPolicy) bool {
-	for _, wc := range pol.WithinClauses {
-		if wc != nil && wc.TriggerOn > 0 {
-			return true
-		}
-	}
-	return false
-}
-
-// withinMatches evaluates temporal trigger clauses against the policy's runtime
-// window.
-// "within N { trigger on M }" — fires when M events happen within N seconds.
-// "within N { trigger until M }" — fires until M events happen within N seconds, then stops.
-//
-// MULTIPLE within clauses are combined with AND (#4423 M3): the policy fires
-// only when EVERY within clause passes for this event. A policy that wants
-// OR-of-conditions must be written as separate policies. This AND semantics is
-// the documented contract (pkg/eventengine/README.md) and is covered by
-// TestWithinMultipleClausesAreANDed.
-//
-// The body is three passes over the clause list rather than one fused loop, so
-// no verdict depends on the order the operator wrote the clauses in — see the
-// #7223 note on pass 2. One deliberate ordering delta survives: a policy that
-// carries BOTH a structurally invalid clause (rejected by pass 1) and a
-// below-threshold trigger-on clause no longer clears the edge latch, where the
-// fused loop did if the below-threshold clause came first. Such a policy fails
-// closed on every event for as long as the invalid clause exists, so the
-// latch's value is unobservable until the config is corrected — after which the
-// next below-threshold event re-arms it through the normal path.
-func (e *Engine) withinMatches(pol *config.EventPolicy, rt *policyRuntime, eventName string, now time.Time) bool {
-	if len(pol.WithinClauses) == 0 {
-		return true // no temporal filter
-	}
-
-	timestamps := rt.windows[eventName]
-
-	// PASS 1 — structural validity, over every clause.
-	//
-	// #3751 defense-in-depth: a within clause with no USABLE positive
-	// threshold (both trigger on/until absent or <= 0) — or a non-positive
-	// window — must fail CLOSED, not always-fire. Commit-time validation
-	// (config.validateEventOptionsWithinAST) rejects such a clause outright,
-	// so this belt only fires on a config that slipped through a TOLERANT
-	// load / peer-sync path (an older binary silently coerced a within/
-	// trigger typo to 0). Falling through to `return true` here would treat
-	// that mis-arrived 0 as an unconditional match — turning a threshold-
-	// gated remediation into an ALWAYS-FIRE one, the original fail-open. A
-	// policy with NO within clauses at all is handled above (return true) —
-	// that legitimately means "no temporal filter"; this guard is only for
-	// a within clause that exists but gates nothing.
-	for _, wc := range pol.WithinClauses {
-		if wc.Seconds <= 0 || (wc.TriggerOn <= 0 && wc.TriggerUntil <= 0) {
-			return false
-		}
-	}
-
-	counts := make([]int, len(pol.WithinClauses))
-	for i, wc := range pol.WithinClauses {
-		window := time.Duration(wc.Seconds) * time.Second
-		for _, ts := range timestamps {
-			if now.Sub(ts) <= window {
-				counts[i]++
-			}
-		}
-	}
-
-	// PASS 2 — the "trigger on N" edge latch, decided over EVERY trigger-on
-	// clause before anything returns.
-	//
-	// "trigger on N" is EDGE-triggered (#3756 M1): it fires on the threshold
-	// CROSSING, not on every event while the level stays at or above N. Junos
-	// `trigger on` re-arms only after the count drops back below N; the 30s
-	// cooldown alone only THROTTLES a sustained level (re-remediating it every
-	// cooldown), which is harmful for a non-idempotent then-batch and spams
-	// commit/rollback history.
-	//
-	// #7223: the re-arm and the latch check are two separate passes
-	// over the clause list, and that separation IS the fix. The single fused
-	// loop this replaces returned false from inside the walk, so with more
-	// than one `within { trigger on N }` clause the outcome depended on the
-	// order the operator wrote them in. Clauses are ANDed (below), so consider
-	// `within 600 { trigger on 10 }` written BEFORE `within 60 { trigger on 3 }`
-	// with the policy already latched: the 600s clause is still at/above 10, so
-	// it hit the latch check and returned — and the 60s clause, which had
-	// dropped below 3 and was the one that should have RE-ARMED the latch, was
-	// never reached. The AND became false and then true again (a real crossing)
-	// and the policy stayed silent until the long window decayed. Written in
-	// the other order the same config re-arms correctly. Deciding the re-arm
-	// across all clauses first makes the result independent of clause order:
-	// the latch clears exactly when the ANDed condition is false, which is
-	// exactly when any trigger-on clause is below its threshold.
-	belowTriggerOn := false
-	hasTriggerOn := false
-	for i, wc := range pol.WithinClauses {
-		if wc.TriggerOn <= 0 {
-			continue
-		}
-		hasTriggerOn = true
-		if counts[i] < wc.TriggerOn {
-			belowTriggerOn = true
-		}
-	}
-	if belowTriggerOn {
-		// Dropped below the threshold: re-arm so the next crossing fires again.
-		rt.onLatched[eventName] = false
-		return false
-	}
-	if hasTriggerOn && rt.onLatched[eventName] {
-		// Level still at/above N on every clause and this crossing already
-		// fired: suppress until some clause's count drops below its threshold
-		// (re-arm above). evaluateEvent sets the latch only AFTER the cooldown
-		// check passes, so a cooldown-suppressed crossing is not consumed.
-		return false
-	}
-
-	// PASS 3 — "trigger until N".
-	//
-	// "trigger until N" fires through the INCLUSIVE N-th event, then stops
-	// (#3756 M2). Junos reads it as "trigger UNTIL the event has been received
-	// N times" — the N-th occurrence is the LAST that fires. The event is
-	// appended to the window BEFORE this check (evaluateEvent), so the N-th
-	// matching event already makes count==N; using `>=` here made count==N
-	// return false, so the N-th never fired and `until 1` (count==1 on the
-	// first event) could NEVER fire at all — a dead-config bug. `>` fires on
-	// 1..N and stops at N+1.
-	for i, wc := range pol.WithinClauses {
-		if wc.TriggerUntil > 0 && counts[i] > wc.TriggerUntil {
-			return false
-		}
-	}
-
-	return true
-}
-
-// pruneWindow removes timestamps older than the policy's maximum within window
-// (or a 60s default). Called on every append so a cooldown-suppressed event
-// can never grow the window unbounded (SMR finding 1).
-func (e *Engine) pruneWindow(pol *config.EventPolicy, eventName string, now time.Time) {
-	rt := e.runtime[pol.Name]
-	if rt == nil {
-		return
-	}
-
-	maxWindow := time.Duration(0)
-	for _, wc := range pol.WithinClauses {
-		w := time.Duration(wc.Seconds) * time.Second
-		if w > maxWindow {
-			maxWindow = w
-		}
-	}
-	if maxWindow == 0 {
-		maxWindow = 60 * time.Second
-	}
-
-	timestamps := rt.windows[eventName]
-	pruned := timestamps[:0]
-	for _, ts := range timestamps {
-		if now.Sub(ts) <= maxWindow {
-			pruned = append(pruned, ts)
-		}
-	}
-	// #4423 M4: the in-place compaction above reuses the SAME backing array, so
-	// its capacity stays pinned at the burst high-water mark forever — a single
-	// storm of a rapidly-flapping probe permanently retains that memory even
-	// after the window drains back to a handful of entries. When the retained
-	// capacity dwarfs the live length, copy into a right-sized slice so the old
-	// backing array can be collected. The threshold keeps steady-state churn
-	// cheap (no realloc for a window oscillating near its typical size).
-	if cap(pruned) >= 64 && cap(pruned) > 4*len(pruned) {
-		shrunk := make([]time.Time, len(pruned))
-		copy(shrunk, pruned)
-		pruned = shrunk
-	}
-	rt.windows[eventName] = pruned
 }
